Plus, minus, times, divided
Solution to the task:    https://putka-upm.acm.si/tasks/t/pmkd/en/
(you are given 6 numbers and a target number C, and can use +, -, * or / operators. You can use all or some of the 6 numbers, and perform the mentioned basic operations, to get as close to the target number C as possible, in the excercise, + and - are done before / and *, furthermore, the output should be the absolute value of the difference between C, and the closest possible value, i.e. the distance between the closest possible value generated and the target C)