Boxes
Solution to the task: https://putka-upm.acm.si/tasks/t/skatle/en/
The boxes must necessarily be nested, e.g., the first inside the second and the second inside the third. WHAT IS THE MAXIMUM NUMBER OF of boxes that can we nest in this way? We are dealing with D-dimensional boxes, and a box X1 fits into box X2, if it holds that X1[i] < X2[i] for all 0 <= i < D, meaning, that the corresponding components of the vector representation of the box are smaller than that of the second box, that has to hold for all components. 
We can perform rotations when trying to nest the boxes (a rotation is simply a permutation of the vector representation of the box.